fix(admin): Sprint 4 P2 — polish (dedup, RU/EN, codemod artifacts)
P2-1: IconButton dedup (3 copies -> 1 shared)
- Created components/admin/IconButton.jsx (33 lines)
- Removed byte-identical copies from AdminAppointments, AdminDoctors,
  AdminPatients; replaced with import. Behavior unchanged.

P2-2: formatCurrency dedup (2 copies -> 1 shared)
- Created utils/formatCurrency.js (Intl ru-RU UZS, null-safe)
- Removed identical copies from AdminDashboard, AdminFinanceOverview
- useUtils.js formatCurrency kept (different impl, different call sites)

P2-3: REPORT_ENDPOINTS dedup (2 copies -> 1 shared)
- Created utils/reportEndpoints.js (12-entry superset + getReportEndpoint)
- Removed local map from ReportGenerator, inline fn from ReportsManager
- ReportsManager now uses superset (5->12 entries)

P2-4: const [, setX] codemod artifacts restored (3 files, 6 slots)
- WebhookManager: showEditModal, showTestModal restored
- DynamicPricingManager: editingRule, editingPackage restored
- DisplayBoardSettings: boards, showBannerForm restored
- Setters were called but values discarded; now reachable for future UI

P2-5: RU/EN mixing -> RU (admin UI is Russian-facing)
- ReportsManager: 7 EN strings -> RU (toast, button titles, aria-labels)
- UnifiedSettings: 'Accent color' -> 'Акцентный цвет' (heading + desc)

Verification:
- vite build: OK, exit 0, ~26s
- vitest run: 515/515 pass (105 test files)
- eslint on 14 files: 0 errors (9 pre-existing warnings)

Sprint 4 of 4 — final sprint. Admin cleanup roadmap complete.
