+++ frontend/src/components/admin/AdminDoctors.jsx
@@ -16,6 +16,7 @@ import {
   Skeleton,
   Select,
 } from '../ui/macos';
+import IconButton from './IconButton';
 import logger from '../../utils/logger';
 
 const departmentOptions = [
@@ -54,27 +55,6 @@ const getDoctorInitials = (doctor) =>
     .slice(0, 2) || 'Д';
 
 const getDepartmentLabel = (department) => departmentLabels[department] || department || 'Не указано';
-
-const IconButton = ({ label, tone = 'default', onClick, children }) => (
-  <Button
-    type="button"
-    variant="ghost"
-    size="small"
-    onClick={onClick}
-    aria-label={label}
-    title={label}
-    className="admin-w-32-h-32-p-0-radius-var-mac-radius-sm-col-dyn" style={{ '--admin-col0': tone === 'danger' ? 'var(--mac-error)' : 'var(--mac-text-secondary)' }}
-  >
-    {children}
-  </Button>
-);
-
-IconButton.propTypes = {
-  children: PropTypes.node.isRequired,
-  label: PropTypes.string.isRequired,
-  onClick: PropTypes.func.isRequired,
-  tone: PropTypes.oneOf(['default', 'danger']),
-};
 
 const AdminDoctors = () => {
   // P-013 fix: shared ConfirmDialog hook (replaces 1 window.confirm() call).
