+++ frontend/src/components/admin/AdminPatients.jsx
@@ -14,6 +14,7 @@ import {
   Skeleton,
   Select,
 } from '../ui/macos';
+import IconButton from './IconButton';
 import logger from '../../utils/logger';
 // P-013 fix: shared ConfirmDialog hook replacing window.confirm() calls.
 import { useConfirm } from '../common/ConfirmDialog';
@@ -84,27 +85,6 @@ const formatAge = (patient, calculateAge) => {
 
   const age = calculateAge(patient.birthDate);
   return Number.isFinite(age) && age >= 0 ? `${age} лет` : 'Не указано';
-};
-
-const IconButton = ({ label, tone = 'default', onClick, children }) => (
-  <Button
-    type="button"
-    variant="ghost"
-    size="small"
-    onClick={onClick}
-    aria-label={label}
-    title={label}
-    className="admin-w-32-h-32-p-0-radius-var-mac-radius-sm-col-dyn" style={{ '--admin-col0': tone === 'danger' ? 'var(--mac-error)' : 'var(--mac-text-secondary)' }}
-  >
-    {children}
-  </Button>
-);
-
-IconButton.propTypes = {
-  children: PropTypes.node.isRequired,
-  label: PropTypes.string.isRequired,
-  onClick: PropTypes.func.isRequired,
-  tone: PropTypes.oneOf(['default', 'danger']),
 };
 
 const AdminPatients = () => {
